requirements.txt
facebook-scribe==2.0
thrift==0.9.2
nose==1.3.6
mock==1.0.1
coverage==3.7.1
